requirements.txt

fastapi
uvicorn
python-multipart
langchain
langchain-openai
chromadb
python-dotenv
tiktoken
pydantic
requests

langchain-community
